# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Legal Copilot API
  version: 1.0.0
  description: |-
    AI-first practice management API for UK law firms.

    ## Authentication
    All endpoints require authentication via Bearer token.

    ## Rate Limiting
    API requests are rate limited to 100 requests per minute.

    ## Pagination
    List endpoints support pagination via `page` and `limit` query parameters.
  contact:
    name: Legal Copilot Support
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Validation failed
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        details:
          type: object
          additionalProperties:
            nullable: true
          description: Additional error details
      required:
        - error
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
...
    TaskPriority:
      type: string
      enum:
        - low
        - medium
        - high
        - urgent
    TaskStatus:
      type: string
      enum:
        - pending
        - in_progress
        - completed
        - cancelled
    TaskAiSource:
      type: string
      nullable: true
      enum:
        - email
        - document
        - matter
        - other
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        matterId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
        description:
          type: string
          nullable: true
        assigneeId:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        createdById:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        dueDate:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        completedAt:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        checklistItems:
          type: array
          nullable: true
          items:
            nullable: true
        tags:
          type: array
          nullable: true
          items:
            nullable: true
        aiGenerated:
          type: boolean
        aiSource:
          $ref: "#/components/schemas/TaskAiSource"
        sourceEntityType:
          type: string
          nullable: true
        sourceEntityId:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
      required:
        - id
        - matterId
        - title
        - description
        - assigneeId
        - createdById
        - priority
        - status
        - dueDate
        - completedAt
        - checklistItems
        - tags
        - aiGenerated
        - aiSource
        - sourceEntityType
        - sourceEntityId
        - createdAt
        - updatedAt
    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - tasks
        - pagination
...
    CreateTaskRequest:
      type: object
      properties:
        matterId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        assigneeId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        priority:
          $ref: "#/components/schemas/TaskPriority"
        dueDate:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        checklistItems:
          type: array
          items:
            nullable: true
        tags:
          type: array
          items:
            nullable: true
      required:
        - matterId
        - title
...
paths:
...
  /api/tasks:
    get:
      summary: List tasks
      description: Retrieve firm tasks with filters
      tags:
        - Tasks
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
            description: Page number (1-indexed)
            example: 1
          required: false
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            description: Items per page (max 100)
            example: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
            format: uuid
            description: Unique identifier in UUID v4 format
            example: 123e4567-e89b-12d3-a456-426614174000
          required: false
          name: matterId
          in: query
        - schema:
            type: string
            format: uuid
            description: Unique identifier in UUID v4 format
            example: 123e4567-e89b-12d3-a456-426614174000
          required: false
          name: assigneeId
          in: query
        - schema:
            $ref: "#/components/schemas/TaskStatus"
          required: false
          name: status
          in: query
        - schema:
            $ref: "#/components/schemas/TaskPriority"
          required: false
          name: priority
          in: query
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create task
      description: Create a new task for a matter
      tags:
        - Tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...